+++ omod/src/main/java/org/openmrs/module/owa/web/controller/OwaRestController.java
@@ -54,17 +54,23 @@ public class OwaRestController {
 	@RequestMapping(value = "/rest/owa/applist", method = RequestMethod.GET)
 	@ResponseBody
 	public List<App> getAppList() {
+            List<App> appList = new ArrayList<>();
+            if(Context.hasPrivilege("Manage OWA")){
 		appManager.reloadApps();
-		return appManager.getApps();
+		appList = appManager.getApps();
+            }
+            return appList;
 	}
 	
 	@RequestMapping(value = "/rest/owa/settings", method = RequestMethod.GET)
         @ResponseBody
         public List<GlobalProperty> getSettings() {
             List<GlobalProperty> owaSettings = new ArrayList<>();
-            owaSettings.add(Context.getAdministrationService().getGlobalPropertyObject("owa.appFolderPath"));
-            owaSettings.add(Context.getAdministrationService().getGlobalPropertyObject("owa.appBaseUrl"));
-            owaSettings.add(Context.getAdministrationService().getGlobalPropertyObject("owa.appStoreUrl"));
+            if(Context.hasPrivilege("Manage OWA")){
+                    owaSettings.add(Context.getAdministrationService().getGlobalPropertyObject("owa.appFolderPath"));
+                    owaSettings.add(Context.getAdministrationService().getGlobalPropertyObject("owa.appBaseUrl"));
+                    owaSettings.add(Context.getAdministrationService().getGlobalPropertyObject("owa.appStoreUrl"));
+            }
             return owaSettings;
         }
 	
@@ -72,10 +78,12 @@ public class OwaRestController {
         @ResponseBody
         public List<GlobalProperty> updateSettings(List<GlobalProperty> settings) {
             List<GlobalProperty> owaSettings = new ArrayList<>();
-            if (null != settings) {
-                for (GlobalProperty gp : settings) {
-                    Context.getAdministrationService().saveGlobalProperty(gp);
-                    owaSettings.add(gp);
+            if(Context.hasPrivilege("Manage OWA")){
+                if (null != settings) {
+                    for (GlobalProperty gp : settings) {
+                        Context.getAdministrationService().saveGlobalProperty(gp);
+                        owaSettings.add(gp);
+                    }
                 }
             }
             return owaSettings;
@@ -84,42 +92,46 @@ public class OwaRestController {
 	@RequestMapping(value = "/rest/owa/addapp", method = RequestMethod.POST)
         @ResponseBody
         public List<App> upload(@RequestParam("file") MultipartFile file, HttpServletRequest request, HttpServletResponse response) throws IOException {
-            String message;
-            HttpSession session = request.getSession();
-            if (!file.isEmpty()) {
-                String fileName = file.getOriginalFilename();
-                File uploadedFile = new File(file.getOriginalFilename());
-                file.transferTo(uploadedFile);
-                try (ZipFile zip = new ZipFile(uploadedFile)) {
-                    if (zip.size() == 0) {
-                        message = messageSourceService.getMessage("owa.blank_zip");
-                        log.warn("Zip file is empty");
-                        session.setAttribute(WebConstants.OPENMRS_ERROR_ATTR, message);
-                        response.sendError(500, message);
-                    }
-                    ZipEntry entry = zip.getEntry("manifest.webapp");
-                    if (entry == null) {
-                        message = messageSourceService.getMessage("owa.manifest_not_found");
-                        log.warn("Manifest file could not be found in app");
+            List<App> appList = new ArrayList<>();
+            if(Context.hasPrivilege("Manage OWA")){
+                String message;
+                HttpSession session = request.getSession();
+                if (!file.isEmpty()) {
+                    String fileName = file.getOriginalFilename();
+                    File uploadedFile = new File(file.getOriginalFilename());
+                    file.transferTo(uploadedFile);
+                    try (ZipFile zip = new ZipFile(uploadedFile)) {
+                        if (zip.size() == 0) {
+                            message = messageSourceService.getMessage("owa.blank_zip");
+                            log.warn("Zip file is empty");
+                            session.setAttribute(WebConstants.OPENMRS_ERROR_ATTR, message);
+                            response.sendError(500, message);
+                        }
+                        ZipEntry entry = zip.getEntry("manifest.webapp");
+                        if (entry == null) {
+                            message = messageSourceService.getMessage("owa.manifest_not_found");
+                            log.warn("Manifest file could not be found in app");
+                            uploadedFile.delete();
+                            session.setAttribute(WebConstants.OPENMRS_ERROR_ATTR, message);
+                            response.sendError(500, message);
+                        } else {
+                            String contextPath = request.getScheme() + "://" + request.getServerName() + ":"
+                                    + request.getServerPort() + request.getContextPath();
+                            appManager.installApp(uploadedFile, fileName, contextPath);
+                            message = messageSourceService.getMessage("owa.app_installed");
+                            session.setAttribute(WebConstants.OPENMRS_MSG_ATTR, message);
+                        }
+                    } catch (Exception e) {
+                        message = messageSourceService.getMessage("owa.not_a_zip");
+                        log.warn("App is not a zip archive");
                         uploadedFile.delete();
                         session.setAttribute(WebConstants.OPENMRS_ERROR_ATTR, message);
                         response.sendError(500, message);
-                    } else {
-                        String contextPath = request.getScheme() + "://" + request.getServerName() + ":"
-                                + request.getServerPort() + request.getContextPath();
-                        appManager.installApp(uploadedFile, fileName, contextPath);
-                        message = messageSourceService.getMessage("owa.app_installed");
-                        session.setAttribute(WebConstants.OPENMRS_MSG_ATTR, message);
                     }
-                } catch (Exception e) {
-                    message = messageSourceService.getMessage("owa.not_a_zip");
-                    log.warn("App is not a zip archive");
-                    uploadedFile.delete();
-                    session.setAttribute(WebConstants.OPENMRS_ERROR_ATTR, message);
-                    response.sendError(500, message);
                 }
+                appManager.reloadApps();
+                appList = appManager.getApps();
             }
-            appManager.reloadApps();
-            return appManager.getApps();
+            return appList;
         }
 }
